Refactor code: split classes into separate files and created main.py
