+++ substitute.py
@@ -1,0 +1,18 @@
+from player import Player
+
+class Substitute(Player):
+    __is_substitute = True
+
+    def __init__(self, name, age, position, salary):
+        super().__init__(name, age, position, salary)
+
+    def set_info(self, name, age, position, salary):
+        super().set_info(name, age, position, salary)
+
+    def add_to_club(self):
+        print('Substitute added to club')
+
+    @staticmethod
+    def print_info(name, age, position, salary):
+        Player.print_info(name, age, position, salary)
+        print('Is Substitute: Yes')
